Final Submission
Contains the code of:
On/Off button on the top left. Clicking the button will darken the screen. Clicking the screen
will lighten the screen up again.
Time and Date in the top centre of the screen. (it should take real time)
Weather icon in a purple box with rounded edges on the top right of the screen ( I believe it
has a shadow too)
Green WIFI icon on the bottom right of the screen.
App should be responsive for a 1O inch screen to a 15.6 inch screen
App should only display in landscape.
Send us APK for us to test. And send us Github for code review